+++ c/image_utils.hpp
@@ -8,13 +8,13 @@ const int IMG_WIDTH = 1024;
 const int FRAME_WIDTH = 25;
 const int FRAME_HEIGHT = 25;
 
-void imageScaler(uint16_t img_orig[IMG_HEIGHT][IMG_WIDTH],
-                 uint16_t img_scaled[IMG_HEIGHT][IMG_WIDTH],
+void imageScaler(uint8_t img_orig[IMG_HEIGHT][IMG_WIDTH],
+                 uint8_t img_scaled[IMG_HEIGHT][IMG_WIDTH],
                  uint16_t src_height,
                  uint16_t src_width,
                  float factor);
 
-void calcIntegralImages(uint16_t i[IMG_HEIGHT][IMG_WIDTH],
+void calcIntegralImages(uint8_t i[IMG_HEIGHT][IMG_WIDTH],
                         int x_start,
                         int y_start,
                         uint64_t ii[FRAME_HEIGHT][FRAME_WIDTH],
